move GroupMember to tailwind
