+++ chat-react2/src/components/Home/Main/ChatInfo/GroupMember.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { MdPersonRemoveAlt1 } from "react-icons/md";
-import "./GroupMember.scss";
 import { useChat } from "@connectycube/use-chat";
 import { ConfirmationAlert } from "@/components/Shared/ConfirmationAlert";
+import Avatar from "@/Components/Shared/Avatar";
 
 export interface GroupMemberProps {
   userId: number;
@@ -22,8 +22,6 @@ const GroupMember: React.FC<GroupMemberProps> = ({ userId, name, avatar }) => {
     removeUsersFromGroupChat([userId]);
   };
 
-  const initials = name.slice(0, 2);
-
   const isChatOwner = selectedDialog.user_id === currentUserId;
   const isAdmin = selectedDialog.user_id === userId;
   const canRemoveUser = isChatOwner && currentUserId !== userId;
@@ -31,27 +29,18 @@ const GroupMember: React.FC<GroupMemberProps> = ({ userId, name, avatar }) => {
   const lastActivityInfo = userId !== currentUserId ? lastActivity[userId] : "";
 
   return (
-    <div className="member">
-      <div className="member__photo-container">
-        {avatar ? (
-          <img className="member__avatar-img" src={avatar} alt="User avatar" />
-        ) : (
-          <div id="background" className="user__no-img">
-            <span className="name">{initials}</span>
-          </div>
-        )}
-      </div>
-      <div className="member__info-container">
-        <div className="member__name-activity">
-          <p>{name}</p>
-          <p className="last__activity">{lastActivityInfo}</p>
+    <div className="flex items-center p-2 gap-2">
+      <Avatar imageUID={avatar} name={name} className="w-[60px] h-[60px]" />
+      <div className="flex items-center justify-between w-full">
+        <div className="flex flex-col items-start">
+          <p className="text-black font-medium">{name}</p>
+          <p className="text-gray-500 text-sm">{lastActivityInfo}</p>
         </div>
-
         {canRemoveUser && (
           <ConfirmationAlert
             triggerChild={
               <MdPersonRemoveAlt1
-                className="remove__user"
+                className="cursor-pointer"
                 color={"#747474"}
                 size={24}
               />
@@ -61,7 +50,7 @@ const GroupMember: React.FC<GroupMemberProps> = ({ userId, name, avatar }) => {
             onConfirm={handleRemoveUser}
           />
         )}
-        {isAdmin && <span>admin</span>}
+        {isAdmin && <span className="text-blue-600 font-bold ml-2">Admin</span>}
       </div>
     </div>
   );
